+++ lib/builtin/markdown/src/markdown_viewer.dart
@@ -67,7 +67,7 @@ class _MarkdownViewerState extends State<MarkdownViewer> {
 
   void _onBack() => _nav?.back();
   void _onForward() => _nav?.forward();
-  void _onPin() => _nav?.pin();
+  void _onPin() => _nav?.togglePin();
   void _onJumpToPin() => _nav?.jumpToPin();
 
   void _onEdit() {